requirements.txt
ultralytics>=8.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyMuPDF>=1.24.0
easyocr>=1.7.0
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
opencv-python>=4.6.0
